[#1216] neofs-cli: Rename `token` subcommand to `bearer`
It works only with the bearer token.
