+++ cmd/neofs-cli/modules/bearer/create.go
@@ -1,4 +1,4 @@
-package token
+package bearer
 
 import (
 	"context"
